feat: refactor client intake profiles and memos routes, handlers, and validations; migrate to dedicated route files
