+++ src/modules/clients/http.ts
@@ -1,32 +1,36 @@
-import * as handlers from '@/modules/clients/handlers';
-import * as routes from '@/modules/clients/routes';
+import { handlers } from '@/modules/clients/handlers';
+import { routes } from '@/modules/clients/routes';
 import { injectAbility } from '@/shared/middleware/inject-ability';
 import { requireAuth } from '@/shared/middleware/requireAuth';
 import { requireOrgMembership } from '@/shared/middleware/requireOrgMembership';
 import { createHonoApp } from '@/shared/router/factory';
 import { registerOpenApiRoutes } from '@/shared/router/openapi-docs';
 
-const clientsApp = createHonoApp();
+const app = createHonoApp();
 
-clientsApp.use('*', requireAuth(), requireOrgMembership(), injectAbility());
+// TODO: migrate clients and memos routes to dedicated sub-apps (split by resource, follow intakeProfileApp pattern)
+const staffApp = createHonoApp();
+staffApp.use('*', requireAuth(), requireOrgMembership(), injectAbility());
 
-// Clients (Note: No POST/create - clients are created via intake or invitation flows)
-clientsApp.openapi(routes.listClientsRoute, handlers.listClientsHandler);
-clientsApp.openapi(routes.getClientRoute, handlers.getClientHandler);
-clientsApp.openapi(routes.updateClientRoute, handlers.updateClientHandler);
-clientsApp.openapi(routes.deleteClientRoute, handlers.deleteClientHandler);
+staffApp.openapi(routes.listClientsRoute, handlers.listClientsHandler);
+staffApp.openapi(routes.getClientRoute, handlers.getClientHandler);
+staffApp.openapi(routes.updateClientRoute, handlers.updateClientHandler);
+staffApp.openapi(routes.deleteClientRoute, handlers.deleteClientHandler);
 
-// Memos
-clientsApp.openapi(routes.listClientMemosRoute, handlers.listClientMemosHandler);
-clientsApp.openapi(routes.createClientMemoRoute, handlers.createClientMemoHandler);
-clientsApp.openapi(routes.updateClientMemoRoute, handlers.updateClientMemoHandler);
-clientsApp.openapi(routes.deleteClientMemoRoute, handlers.deleteClientMemoHandler);
+staffApp.openapi(routes.listClientMemosRoute, handlers.listClientMemosHandler);
+staffApp.openapi(routes.createClientMemoRoute, handlers.createClientMemoHandler);
+staffApp.openapi(routes.updateClientMemoRoute, handlers.updateClientMemoHandler);
+staffApp.openapi(routes.deleteClientMemoRoute, handlers.deleteClientMemoHandler);
 
-// Intake profile
-clientsApp.openapi(routes.getClientIntakeProfileRoute, handlers.getClientIntakeProfileHandler);
-clientsApp.openapi(routes.updateClientIntakeProfileRoute, handlers.updateClientIntakeProfileHandler);
+const intakeProfileApp = createHonoApp();
+intakeProfileApp.use('*', requireAuth(), requireOrgMembership(), injectAbility());
 
-// Register routes for OpenAPI documentation extraction
-registerOpenApiRoutes(clientsApp, routes);
+intakeProfileApp.openapi(routes.getClientIntakeProfileRoute, handlers.getClientIntakeProfileHandler);
+intakeProfileApp.openapi(routes.updateClientIntakeProfileRoute, handlers.updateClientIntakeProfileHandler);
 
-export default clientsApp;
+app.route('/', staffApp);
+app.route('/intake-profile', intakeProfileApp);
+
+registerOpenApiRoutes(app, routes);
+
+export default app;
